refactor: combine schema move to util folder
